baselines/kiran-meta-ood/meta-ood/requirements.txt
cached-property==1.5.2
Cython==0.29.21
dataclasses
future==0.18.2
h5py==3.1.0
joblib==0.17.0
numpy
Pillow
scikit-learn==0.23.2
scipy
threadpoolctl==2.1.0
torch
torchvision
typing-extensions
pycocotools
opencv-python
scikit-image
git+https://github.com/cocodataset/panopticapi.git
